Replaced 'int' type with 'intptr_t' type in pointer arithmetic for compiling without -fpermisive

+++ MemoryFree.cpp
@@ -4,6 +4,8 @@
 #include <WProgram.h>
 #endif
 
+#include <stdint.h>
+
 extern unsigned int __heap_start;
 extern void *__brkval;
 
@@ -39,13 +41,13 @@ int freeListSize()
 int freeMemory()
 {
   int free_memory;
-  if ((int)__brkval == 0)
+  if ((intptr_t)__brkval == 0)
   {
-    free_memory = ((int)&free_memory) - ((int)&__heap_start);
+    free_memory = ((intptr_t)&free_memory) - ((intptr_t)&__heap_start);
   }
   else
   {
-    free_memory = ((int)&free_memory) - ((int)__brkval);
+    free_memory = ((intptr_t)&free_memory) - ((intptr_t)__brkval);
     free_memory += freeListSize();
   }
   return free_memory;
